requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via flask, flask-jsontools-slippers
flask-jsontools-slippers==0.1.5
flask-sqlalchemy==2.2
flask==0.12.1             # via flask-jsontools-slippers, flask-sqlalchemy
itsdangerous==0.24        # via flask, flask-jsontools-slippers
jinja2==2.9.6             # via flask, flask-jsontools-slippers
markovify==0.6.0
markupsafe==1.0           # via flask-jsontools-slippers, jinja2
six==1.10.0               # via sqlalchemy-utils
sqlalchemy-utils==0.32.14
sqlalchemy==1.1.10
textstat==0.3.1
unidecode==0.4.20         # via markovify
werkzeug==0.12.1          # via flask, flask-jsontools-slippers
